sumé tempState además de acomodar lo de Cloudinary

+++ backend/controllers/imageController.js
@@ -2,20 +2,25 @@ import { uploadImage } from "../libs/cloudinary.js"
 import fs from "fs-extra"
 
 const postBookImage = async (req, res) => {
-  // const formatos = ["png", "jpg", "webp", "gif"]
-  // if (
-  //   !formatos.includes(
-  //     req.files.image.name.split(".")[
-  //     req.files.image.name.split(".").length - 1
-  //     ]
-  //   )
-  // ) {
-  //   return res.status(400).send({ msg: "Invalid image format (jpg, png, webp or gif)" })
-  // }
+  const formatos = ["png", "jpg", "webp", "gif"]
+  if (
+    !formatos.includes(
+      req.files.image.name.split(".")[
+      req.files.image.name.split(".").length - 1
+      ]
+    )
+  ) {
+    return res.status(400).send({ msg: "Invalid image format (jpg, png, webp or gif)" })
+  }
 
   try {
     if (req.files.image) {
-      const response = await uploadImage(req.files.image.tempFilePath)
+      // const response = await uploadImage(req.files.image.tempFilePath) // comentar y descomentar lo sgte p/no subir al pepe ---v
+      const response = {
+        secure_url: 'dummy_secure_url',
+        public_id: 'dummy_public_id'
+      }
+      // // ----------------------------
       console.log("🚀 ~ file: imageController.js ~ line 19 ~ postBookImage ~ response", response)
       await fs.remove(req.files.image.tempFilePath)
